corrigiendo errores, agregando para cerrar sesion

+++ api/routes/templates.js
@@ -3,6 +3,7 @@ const router = express.Router();
 const { checkAuth } = require('../middlewares/authentication');
 //models
 import Template from '../models/template';
+import Device from '../models/device';
 
 // response
 const response = {
@@ -58,6 +59,14 @@ router.delete('/template',checkAuth, async (req, res) => {
   try {
     const userId=req.user._id;
     const templateId = req.query.templateId;
+    const devices = await Device.find({
+      templateId: templateId,
+      userId: userId
+    });
+    if (devices.length > 0) {
+      response.message = "fail";
+      return res.json(response)
+    }
     const result=await Template.deleteOne({
       userId: userId,
       _id: templateId
